requirements.txt
flask[async]
flask_mail
flask_sqlalchemy
flask_login
pefile
ssdeep
scikit-learn
pyexifinfo
whois
beautifulsoup4
peutils
virustotal_python
yara-python
python-whois
python-magic
pandas
pydig
selenium
pyppeteer
